+++ espnff/__init__.py
@@ -1,1 +1,3 @@
-from .espnff import League
+from .espnff import League, Team
+
+__all__=['League', 'Team']